初始化提交代码

+++ src/main/java/com/doctor/visit/domain/BusArticle.java
@@ -7,7 +7,7 @@ import javax.persistence.*;
 @Table(name = "bus_article")
 public class BusArticle implements Serializable {
     @Id
-    private Integer id;
+    private Long id;
 
     /**
      * 文章名称
@@ -18,7 +18,7 @@ public class BusArticle implements Serializable {
      * 文章类型id
      */
     @Column(name = "class_id")
-    private Integer classId;
+    private Long classId;
 
     /**
      * 文章类型名称
@@ -35,7 +35,7 @@ public class BusArticle implements Serializable {
      * 创建者id
      */
     @Column(name = "create_by")
-    private Integer createBy;
+    private Long createBy;
 
     /**
      * 创建者姓名
@@ -53,7 +53,7 @@ public class BusArticle implements Serializable {
      * 修改人
      */
     @Column(name = "edit_by")
-    private Integer editBy;
+    private Long editBy;
 
     /**
      * 修改人姓名
@@ -78,14 +78,14 @@ public class BusArticle implements Serializable {
     /**
      * @return id
      */
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
     /**
      * @param id
      */
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -112,7 +112,7 @@ public class BusArticle implements Serializable {
      *
      * @return class_id - 文章类型id
      */
-    public Integer getClassId() {
+    public Long getClassId() {
         return classId;
     }
 
@@ -121,7 +121,7 @@ public class BusArticle implements Serializable {
      *
      * @param classId 文章类型id
      */
-    public void setClassId(Integer classId) {
+    public void setClassId(Long classId) {
         this.classId = classId;
     }
 
@@ -166,7 +166,7 @@ public class BusArticle implements Serializable {
      *
      * @return create_by - 创建者id
      */
-    public Integer getCreateBy() {
+    public Long getCreateBy() {
         return createBy;
     }
 
@@ -175,7 +175,7 @@ public class BusArticle implements Serializable {
      *
      * @param createBy 创建者id
      */
-    public void setCreateBy(Integer createBy) {
+    public void setCreateBy(Long createBy) {
         this.createBy = createBy;
     }
 
@@ -220,7 +220,7 @@ public class BusArticle implements Serializable {
      *
      * @return edit_by - 修改人
      */
-    public Integer getEditBy() {
+    public Long getEditBy() {
         return editBy;
     }
 
@@ -229,7 +229,7 @@ public class BusArticle implements Serializable {
      *
      * @param editBy 修改人
      */
-    public void setEditBy(Integer editBy) {
+    public void setEditBy(Long editBy) {
         this.editBy = editBy;
     }
