+++ src/main/java/com/doctor/visit/domain/BusArticleClass.java
@@ -7,7 +7,7 @@ import javax.persistence.*;
 @Table(name = "bus_article_class")
 public class BusArticleClass implements Serializable {
     @Id
-    private Integer id;
+    private Long id;
 
     /**
      * 文章分类名称
@@ -18,7 +18,7 @@ public class BusArticleClass implements Serializable {
      * 创建者id
      */
     @Column(name = "create_by")
-    private Integer createBy;
+    private Long createBy;
 
     /**
      * 创建者姓名
@@ -36,7 +36,7 @@ public class BusArticleClass implements Serializable {
      * 修改人
      */
     @Column(name = "edit_by")
-    private Integer editBy;
+    private Long editBy;
 
     /**
      * 修改人姓名
@@ -61,14 +61,14 @@ public class BusArticleClass implements Serializable {
     /**
      * @return id
      */
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
     /**
      * @param id
      */
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -95,7 +95,7 @@ public class BusArticleClass implements Serializable {
      *
      * @return create_by - 创建者id
      */
-    public Integer getCreateBy() {
+    public Long getCreateBy() {
         return createBy;
     }
 
@@ -104,7 +104,7 @@ public class BusArticleClass implements Serializable {
      *
      * @param createBy 创建者id
      */
-    public void setCreateBy(Integer createBy) {
+    public void setCreateBy(Long createBy) {
         this.createBy = createBy;
     }
 
@@ -149,7 +149,7 @@ public class BusArticleClass implements Serializable {
      *
      * @return edit_by - 修改人
      */
-    public Integer getEditBy() {
+    public Long getEditBy() {
         return editBy;
     }
 
@@ -158,7 +158,7 @@ public class BusArticleClass implements Serializable {
      *
      * @param editBy 修改人
      */
-    public void setEditBy(Integer editBy) {
+    public void setEditBy(Long editBy) {
         this.editBy = editBy;
     }
