+++ src/main/java/com/doctor/visit/domain/BusClincClass.java
@@ -10,7 +10,7 @@ public class BusClincClass implements Serializable {
      * 主键
      */
     @Id
-    private Integer id;
+    private Long id;
 
     /**
      * 门诊名称
@@ -28,7 +28,7 @@ public class BusClincClass implements Serializable {
      * 创建者id
      */
     @Column(name = "create_by")
-    private Integer createBy;
+    private Long createBy;
 
     /**
      * 创建者姓名
@@ -46,7 +46,7 @@ public class BusClincClass implements Serializable {
      * 修改人
      */
     @Column(name = "edit_by")
-    private Integer editBy;
+    private Long editBy;
 
     /**
      * 修改人姓名
@@ -73,7 +73,7 @@ public class BusClincClass implements Serializable {
      *
      * @return id - 主键
      */
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -82,7 +82,7 @@ public class BusClincClass implements Serializable {
      *
      * @param id 主键
      */
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -127,7 +127,7 @@ public class BusClincClass implements Serializable {
      *
      * @return create_by - 创建者id
      */
-    public Integer getCreateBy() {
+    public Long getCreateBy() {
         return createBy;
     }
 
@@ -136,7 +136,7 @@ public class BusClincClass implements Serializable {
      *
      * @param createBy 创建者id
      */
-    public void setCreateBy(Integer createBy) {
+    public void setCreateBy(Long createBy) {
         this.createBy = createBy;
     }
 
@@ -181,7 +181,7 @@ public class BusClincClass implements Serializable {
      *
      * @return edit_by - 修改人
      */
-    public Integer getEditBy() {
+    public Long getEditBy() {
         return editBy;
     }
 
@@ -190,7 +190,7 @@ public class BusClincClass implements Serializable {
      *
      * @param editBy 修改人
      */
-    public void setEditBy(Integer editBy) {
+    public void setEditBy(Long editBy) {
         this.editBy = editBy;
     }
